Seg.Net Refactor and Model Lib.
Refactor made these classes final static classes and this solved one memory issue, but also means that the memory use is very high as soon as the app boots up and stores the high res model.  I'm debating adding a memory usage view inside the app that will help the user understand why it is so difficult for the app to run without crashing.