MenuPickerBugs
The menu picker has a strange bug where when I close a sheet view in the gallery after a source image is already selected, the menu doesnt popup anymore, instead the menu simply highlights blue when pressed.  This is a SwiftUI bug and using the wheel picker works without this bug.